Refactor routes, get_path: use strategy patterns
We had factored the graph providers and search algorithms to use
strategy design patterns, but we were not really using the power of
abstraction in get_route (we were depending on concretions instead of
abstractions). Now all of the dependencies on concrete search algos
and graph providers are localized to routes.py.